moved hostview and finishedButtons components to be resuable
