+++ client/components/quiz/HostFinished.js
@@ -2,7 +2,7 @@ import React from 'react'
 import {connect} from 'react-redux'
 import database from '../../firebase'
 import {startGameThunk, resetGameThunk, deleteGameThunk} from '../../store/game'
-import FinishedButtons from './FinishedButtons'
+import FinishedButtons from '../FinishedButtons'
 import {
   BarChart,
   Bar,
